mci: sdhci: use the busy_timeout value in the sdhci_wait_idle functions
The busy_timeout value is used by the sdhci_wait_idle and
sdhci_wait_idle_data functions to define the timeout to wait on when
send a command to the controller.

The default value remains unchanged and is set to 10ms.

In pbl is_timeout() is defined as:

With this patch the timeout argument to wait_on_timeout() is no longer a
constant. wait_on_timeout() is implemented using is_timeout(). The above
division must now be done at runtime. On 32bit ARM linking fails with an
undefined reference to __aeabi_uldivmod. Work around this by replacing
the division with a 10 bit rightshift resulting in a division by 1024
which is close enough to the desired value of 1000.

Signed-off-by: Sebastien Bourdelin <sebastien.bourdelin@rtone.fr>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241106153702.858977-4-sebastien.bourdelin@rtone.fr
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>