ARM: Rockchip: implement tee.bin v1 header parsing
Building OP-TEE produces a number of binaries among them tee.elf,
tee.bin and tee-raw.bin.

So far, we assumed in Rockchip code that the binary provided by the user
to be built into barebox is tee-raw.bin.

That binary is installed as-is to address 0x200000, which is not
generally applicable:

  - At least in upstream OP-TEE, the upstream RK3399 support uses
    0x30000000 as base address

  - Even if OP-TEE didn't hang, 0x200000 is likely not at the end
    of RAM, but that's where barebox assumes OP-TEE is by default,
    and thus barebox would carve out unneeded reserved memory.

The tee.bin on the other hand has an OP-TEE header in front of it, which
our i.MX OP-TEE support already uses to identify where to load the binary
too and how much memory to reserve.

Let's do the same for Rockchip and check if have an OP-TEE header and
use that to inform our placement of OP-TEE and reservation of memory in
barebox itself and in the reserved memory entries that will be fixed up
for Linux.

As OP-TEE support is added anew, we also make sure to always have
a OP-TEE header for barebox proper to consult and we fake one if
there isn't any.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241106185724.2526-6-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>