ARM: Rockchip: use vendor blob OPTEE load addresses
It's unknown where the 0x200000 came from, but it's most likely wrong.

 - For the RK3399, the upstream default is 0x30000000 and
   the vendor load address in rkbin/RKTRUST/RK3399TRUST.ini is
   0x08400000 for the blob.

 - For the RK3568, there's no upstream support and for both RK35{68,88}
   the vendor load address in rkbin/RKTRUST/RK356xTRUST.ini is
   0x08400000 for the blob.

The hardcoded OP-TEE load address in barebox is only relevant, when the
OP-TEE header is missing. Users compiling OP-TEE themselves can use the
tee.bin and ignore the hardcoded value altogether.

Users using the blob don't have the header, so it's a good trade off to
use the vendor load addresses as default addresses.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241106185724.2526-7-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>