fs: tftp: cache dentries for some time
For TFTP we always returned invalid in the d_revalidate op. This has
the effect that we always look on the server if a file still exists, but
has also the effect that we initiate multiple tftp connections when we
want to transfer a file.

This patch introduces a grace time of two seconds in which a dentry will
be valid before looking it up on the server again. Two seconds is long
enough for not initiating multiple TFTP connections when transferring a
file and also short enough that changes on the server will be recognized
by barebox.

Link: https://lore.kernel.org/r/20241107-fs-tftp-v1-2-39d7e1cff418@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>