dma: add dma-devices support
This adds support for dma-devices (on Linux aka dmaengines) to barebox.
The code is based on U-Boot-2025.01-rc1.

The code consists of a shim wrapper layer around a struct dma_ops and
helper functions to retrieve the suitable dma_ops instance from the
devices tree based on standard dma properties.

Link: https://lore.kernel.org/r/20241108-network-k3-v1-4-ee71bff15eb7@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>