watchdog: cleanup timeout_cur parameter
Currently a watchdog has a timeout_cur device parameter which has
a slightly confusing behaviour. The parameter has nothing to do with the
currently configured watchdog timeout, instead it is merely used by
the automatic watchdog feeder which configures the watchdog timeout
with that value.

Refine the semantics of the parameter by using it as the watchdog
timeout as the name suggests. When read, it returns the currently
configured timeout. When set, it configures the watchdog timeout
and starts it (unless it's set to 0 in which case the watchdog will
be disabled).

Link: https://lore.kernel.org/r/20241112-k3-watchdog-v1-1-ceb254779735@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>