watchdog: add watchdog ping support
The barebox watchdog API is quite limited. the set_timeout() hook is an
all purpose function for enabling and disabling the watchdog and also to
configure the timeout.
This is not suitable for watchdogs which for example can't change the
timeout anymore once started.
This patch introduces a watchdog_ping() function which hooks into the
drivers with a struct watchdog::ping function. With this the watchdog
can be pinged without caring about the timeout.  The ping hook is
optional. When not implemented watchdog_ping() falls back to call
watchdog_set_timeout() with the current watchdog timeout.

Link: https://lore.kernel.org/r/20241112-k3-watchdog-v1-2-ceb254779735@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>