net: davinci_mdio: drop phy_mask detection
The davinci_mdio driver tries to autodetect connected phys using the
controllers internal state engine that seems to periodically scan for
phys. It then sets phy_mask according to the phys found. This doesn't
work on a beagleplay board because the timeout is too short.

This phy detection mechanism is an optimization only, which in the best
case saves some time needed to scan the bus for phys later on. Normally
the phys are described in the device tree, in this case the bus isn't
scanned for devices anyway, so instead of adjusting the timeout we can
drop the whole thing altogether.

Link: https://lore.kernel.org/r/20241113-davinci-mdio-clk-fix-v1-1-6a98c9864bf6@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>