net: davinci_mdio: initialize divider from clk
Instead of using a fixed divider value calculate a divider from the
input clock.

We currently do not have clk support for am33xx, so we can't handle
no clk as an error. In this case assume a clock frequency of 256MHz.
With a default desired MDIO clock of 1MHz (used in many TI dtsi files)
we get a divider value of 0xff which we previously used as hardcoded
value.

Link: https://lore.kernel.org/r/20241113-davinci-mdio-clk-fix-v1-3-6a98c9864bf6@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>