remoteproc: imx: add support for loading ELF data sections into DRAM
The Cortex-A MPU and the Cortex-M co-processor on the i.MX have
different address space layouts. The driver thus keeps
struct imx_rproc_att[] tables for the different SoCs to translate device
addresses in the ELF headers to the virtual addresses used by the
Cortex-A, which loads the ELF image.

The maximum size of this mapping was so far 8, which may not be large
enough for some ELF images. Enlarge it to 32 like Linux does.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241119085641.2598035-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>