ARM: dts: skov-imx6: switch barebox' environment from MBR to GPT
When barebox is booted from an SD-card the latter is usually not MBR-partitioned
anymore but instead utilizes a GUID partition table today which assigns names to
partitions. This allows for more flexibility as it is independent of a specific
partition number. Chose the name "barebox-environment" for the partition that
shall be used for storing barebox' environment.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>
Link: https://lore.kernel.org/r/20241119113328.2024731-1-u.oelmann@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>