usb: typec: fix freeing uninitialized memory
When the code was first added, a kzalloc was allocating the memory for
the struct device. This has been recently changed to malloc as a
shifting of the header caused the function to be no longer defined.

malloc() doesn't zero the memory though, leading to a potential freeing
of uninitialized memory in dev_set_name.

Fix this by using calloc() instead. Reintroducing kzalloc() would be
fine too, but calloc() saves some space, because it guarantees a smaller
alignment.

Fixes: d9fa8b26b60c ("usb: typec: fix compile test with CONFIG_SANDBOX")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241122095326.2877232-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>