ARM: omap: add two smaller MLO defconfigs for AM335x
The Beaglebone Black remains a popular board and it's not a good look to
have it not compile by default due to barebox getting too big.

As we keep running into size issues with the MLO defconfig, let's accept
that we die bite off more than we could chew and split off the defconfigs
into one for SD/MMC and another for NOR/NAND and document that users
should prefer those instead.

Reported-by: Sohaib Mohamed <sohaib.amhmd@gmail.com>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241202080818.3471605-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>