checkpatch: fix detection of barebox root directory
In v2024.09.0, we switched from our old README to a new fancy
README.rst. What was missing though, was adjusting scripts/checkpach.pl
as it uses the existence of a number of files, including the README file
to decide whether a directory is the top-level barebox directory or not.

Rename README to README.rst and while at it, change the error messages
to not output a confusing:

  checkpatch.pl: Must be run from the top-level dir. of a kernel tree

Reported-by: danct12 # IRC
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241202124232.1813897-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>