ARM: rockchip: Add support for Pine64 PineTab 2
The Pine64 PineTab 2 tablet is basically the Quartz64 but as
a finished product.

There are (currently) two revisions, v0.1 and v2.0.

v0.1 was sent to developers and there are a few units around
(probably less than 10?)

v2.0 is the consumer available version, this version changed the
display reset pin and hooked WLAN/BT chip to a GPIO.

Signed-off-by: Dang Huynh <danct12@riseup.net>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241203-pt2-init-v2-1-d6c625e1e455@riseup.net
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>