ARM: psci-client: better search for existing psci node
The PSCI fixup code assumes the psci node is at /psci. This is not the
case for all boards. TI AM62x upstream dtsi files have the psci node at
/firmware/psci where the current code doesn't find it and creates a new
node, so on this SoC we end up with two PSCI nodes.

Fix this by searching for a node compatible to "arm,psci-0.2" or
"arm,psci-1.0" instead of assuming a fixed location.

Link: https://lore.kernel.org/r/20241209103025.359093-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>