fdt: Add function to check if a pointer contains a fdt
When barebox is started 2nd stage from another barebox or U-Boot it
usually gets a device tree passed in a register. We can't rely on this
though, so before assuming a pointer has a device tree we need to
perform some basic checks for being in a certain range, is sufficiently
aligned and actually contains a device tree.
This adds a function performing these checks.

Link: https://lore.kernel.org/r/20241209-ext-dt-handoff-v1-1-e61052d5d437@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>