video: mipi_dbi: don't send image before initialization
Don't send an image to the display before it is initialized, as it can
cause a broken image being stored in the display memory and being shown
while initializing. Instead just damage the image in the flush call but
avoid writing the actual image.

This problem was experienced with a ILI9341 LCD driver and calling the
splash command before enabling the framebuffer. The splash command flushed
out the data, but given no initialization, they were misinterpreted.
After enabling the display, it displayed the broken image for a short time
before being updated. This caused a flickering effect for the viewer.
Switching the command ordering also fixes the problem, but sends an black
image before the actual splash image.

Co-developed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Link: https://lore.kernel.org/r/20241210080403.1235658-1-s.pueschel@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>