net: tap: return -ENODEV when TAP permission is denied
tap_alloc will return -EPERM, when not running with CAP_NET_ADMIN.
For most users, this is the default case and thus barebox printing:

  could not get tap device: Operation not permitted
  ERROR: tap tap0: probe failed: Operation not permitted

is just confusing. Improve this by returning -ENODEV in that situation
instead, which driver core already understands to be an acceptable
return code that warrants no error message.

The user will then be shown only the first message with no scary red
ERROR prefix.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241212072228.2468741-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>