ARM: beaglebone: register bbu handlers for FAT partition
We currently only register a single barebox update handler for the eMMC,
but the MLO can also be on the SD-Card or in a FAT partition instead one
of the magic offsets.

Let's therefore register MLO bbu handlers for those as well as well as
for the second stage barebox.bin.

Fixes: df22a22f84b6 ("bbu: fix exporting i.MX NAND bbu handler over fastboot")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241212075308.2499846-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>