environment: don't return raw device when file-path is specified
Environment nodes that specify the file-path property expect the
environment to be stored in a file system.

Returning the raw device path when CONFIG_OF_BAREBOX_ENV_IN_FS is
disabled in such a case is not a good idea, as saving the env may
then unexpectedly overwrite unrelated device content. Bail out in
case a file-path is specified, but we have no env in FS support.

Signed-off-by: Lucas Stach <dev@lynxeye.de>
Link: https://lore.kernel.org/r/20241213150855.263388-1-dev@lynxeye.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>