optee: fix i.MX8M link error when loading OP-TEE without CONFIG_TEE
CONFIG_TEE is not required for barebox to load OP-TEE, just for
communicating with it.

The Kconfig option for OF_FIXUP_OPTEE, which guards a symbol that is
used by imx8_soc_imx8m_init() whenever CONFIG_PBL_OPTEE is defined is
made conditional on CONFIG_TEE though, leading to alink error whenever
CONFIG_TEE is undefined:

  ld: ./drivers/soc/imx/soc-imx8m.c:219: undefined reference to
      `of_optee_fixup'

Fix this by making the symbol not depend on CONFIG_TEE again.

Fixes: 56535d9cb0fc ("optee: make OP-TEE OF fixup ARM-specific")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241214084834.959213-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>