i2c: i2c: fix getting adapter.nr for I2C muxes
When probed from device tree, "real" I2C controllers are virtual
devices that have the OF device as parent.

For those the current logic of using the alias of the parent
device's of_node works.

I2C muxes are different, because their parent device is the virtual
device of their controller. Using that controller's alias is often
destined to result in an -EBUSY:

  pca9450-i2c pca9450b0: PMIC Chip ID: 0x3
  ERROR: i2c1: failed to add mux-adapter (error=-16)
  ERROR: pca954x pca95450: failed to register multiplexed adapter0

As a single I2C mux can have multiple channels, there may also be just
one OF device for multiple virtual adapters. As all I2C adapters being
registered in barebox also have the of_node of the virtual device
populated, let's just consult that.

Fixes: b4746725454c ("i2c: get adapter.nr from device tree")
Reported-by: Alexander Shiyan <eagle.alexander923@gmail.com>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241215095412.3756337-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>