spi: add support to setup spi-cs-{setup,hold,inactive}-delay-ns
Add support to parse and setup the common OF properties. The parsing is
quite confusing since the kernel driver API decided to use a u16 for the
value. I kept it this way to be closer to the Linux code.

This prepares the core for the upcoming core based message handling.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.kernel.org/r/20241216130324.1592755-7-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>