commands: addpart: allow force creation of overlapping partition
While we have the guard rails for good reason, a user on the shell
should be able to override them and create partitions as they see fit,
e.g. if they have a MBR they don't care about anyway, it should still
be possible to force create a new fixed partition that overlaps it
and pass that to Fastboot for example.

Let's add a new -f (force) parameter that enables this.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250106090048.3420243-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>