include: rename printk.h to printf.h
We also have <linux/printk.h>, which defines the pr_* and dev_* family
of Linux logging functions. printk.h only defines the base printing
functions as well as the printk macro, which expands to printf.

To make it clearer, which is which, rename printk.h to printf.h.

No functional change.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250106091906.3435166-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>