driver: replace single info callbacks with stack
A number of driver register an info callback that prints
driver-specific information for devices that already have a device
callback. This is currently handled in every individual driver by
storing the old info callback in device-specific driver data.

Move this functionality into the driver core by means of new
a devinfo_add() function.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250106091915.3435362-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>